+++ packages/q3-admin/src/components/Search/Search.jsx
@@ -23,7 +23,11 @@ export const Search = () => {
 
   const handleKeyCode = (e) => {
     const val = e.target.value;
-    if (e?.code !== 'Enter') return;
+    if (
+      !['Enter', 'NumpadEnter'].includes(e?.code) &&
+      e?.key !== 'Enter'
+    )
+      return;
 
     navigate(
       val