+++ packages/q3-admin/src/containers/Notifications/Notifications.jsx
@@ -1,56 +1,17 @@
 import React from 'react';
 import Notifications from 'q3-ui-notifications';
-import { array } from 'q3-ui-helpers';
-import { SocketContext } from '../Socket';
+import { useNotifications } from '../../hooks';
 
 const NotificationsContainer = () => {
-  const [list, setList] = React.useState([]);
-  const { emit, on } = React.useContext(SocketContext);
-
-  const dataToListState = ({ data }) =>
-    setList((prevState = []) =>
-      (array.hasLength(prevState)
-        ? [].concat(data).concat(prevState)
-        : [data]
-      )
-        .flat()
-        .filter(Boolean)
-        .map((item) => ({
-          label: item.path,
-          ...item,
-        })),
-    );
-
-  const sendToSocket = (eventName) => (
-    eventInstance,
-    id,
-  ) => {
-    emit(eventName, id, () => {
-      setList((prev) =>
-        prev.map((item) => ({
-          ...item,
-          ...(item.id === id
-            ? {
-                'hasDownloaded': true,
-                'hasSeen': true,
-              }
-            : {}),
-        })),
-      );
-    });
-  };
-
-  React.useEffect(() => {
-    on('download', dataToListState);
-    on('recent', dataToListState);
-  }, []);
+  const { acknowledge, data, error } = useNotifications();
 
   return (
     <Notifications
-      data={list}
+      data={data}
+      error={error}
       // we'll handle both the same way for now
-      onClick={sendToSocket('acknowledge')}
-      onView={sendToSocket('acknowledge')}
+      onClick={acknowledge}
+      onView={acknowledge}
     />
   );
 };