+++ packages/q3-admin/src/containers/index.js
@@ -6,7 +6,6 @@ export { default as Header } from './header';
 export { default as Table } from './table';
 export { default as Page } from './page';
 export { default as SubDetail } from './subDetail';
-export { MongoChart, MongoStats } from './visualization';
 
 export * from './state';
 export { default as connect } from './connect';