+++ packages/q3-admin/src/hooks/useRefresh.test.js
@@ -1,5 +1,10 @@
 import React from 'react';
 import useRefresh from './useRefresh';
+import { invokeDocumentListener } from './useNotifications';
+
+jest.mock('./useNotifications', () => ({
+  invokeDocumentListener: jest.fn(),
+}));
 
 jest.mock('q3-ui-helpers', () => ({
   browser: {
@@ -13,87 +18,33 @@ jest.mock('q3-ui-helpers', () => ({
   },
 }));
 
-let useContext;
-let useContextReturnFn;
-
-let join;
-let watch;
-let leave;
-
-beforeAll(() => {
-  window.location.search = '?sort=name';
-});
-
 beforeEach(() => {
-  join = jest.fn();
-  watch = jest.fn();
-  leave = jest.fn();
-
   jest
     .spyOn(React, 'useEffect')
     .mockImplementation((fn) => {
-      useContextReturnFn = fn();
+      fn();
     });
-
-  useContext = jest.spyOn(React, 'useContext');
 });
 
 describe('useRefresh', () => {
-  it('should skip when given an id', () => {
-    useContext.mockReturnValue({
-      join,
-      id: 1,
+  it('should debounce poll', (done) => {
+    jest.spyOn(React, 'useContext').mockReturnValue({
+      collectionName: 'test',
     });
 
-    useRefresh();
-    expect(join).not.toHaveBeenCalled();
-  });
-
-  it('should join and leave the room', () => {
-    const collectionName = 'test';
-
-    const shouldHaveBeenCalledWithCollection = (fn) =>
-      expect(fn).toHaveBeenCalledWith({ collectionName });
-
-    useContext.mockReturnValue({
-      collectionName,
-      join,
-      leave,
-      watch,
-    });
-
-    useRefresh();
-    useContextReturnFn();
-    shouldHaveBeenCalledWithCollection(join);
-    shouldHaveBeenCalledWithCollection(leave);
-  });
-
-  it('should poll', (done) => {
     const poll = jest.fn().mockResolvedValue({});
-    useContext.mockReturnValue({
-      join,
-      watch,
-    });
 
-    useRefresh(poll, 9);
-    watch.mock.calls[0][0]('3');
+    invokeDocumentListener.mockImplementation(
+      (eventName, fn) => {
+        Array.from({ length: 10 }).forEach(() => fn());
+      },
+    );
 
-    setTimeout(() => {
-      expect(poll).toHaveBeenCalledWith('?sort=name');
-      done();
-    }, 10);
-  });
+    useRefresh(poll, 8);
 
-  it('should debounce', (done) => {
-    const poll = jest.fn().mockResolvedValue({});
-    useContext.mockReturnValue({
-      join,
-      watch,
-    });
-
-    useRefresh(poll, 9);
-    Array.from({ length: 10 }).forEach(() =>
-      watch.mock.calls[0][0](),
+    expect(invokeDocumentListener).toHaveBeenCalledWith(
+      'q3-change-stream-test',
+      expect.any(Function),
     );
 
     setTimeout(() => {