+++ packages/q3-ui-charts/src/components/TooltipList/useStyle.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles(() => ({
+  root: {
+    listStyle: 'none',
+  },
+  item: {
+    margin: 0,
+    fontSize: '0.933rem',
+  },
+}));