+++ packages/q3-ui/src/list/components/listItem/listItem.jsx
@@ -1,11 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { Link } from '@reach/router';
 import { navigate } from '@reach/router';
+import Box from '@material-ui/core/Box';
 import ListItemMui from '@material-ui/core/ListItem';
 import ListItemTextMui from '@material-ui/core/ListItemText';
 import ListItemAvatarMui from '@material-ui/core/ListItemAvatar';
 import Avatar from '@material-ui/core/Avatar';
 import { formatArrayAsCommaDelineatedString } from '../../utils';
+import useStyles from './useStyles';
 
 export const ListItem = ({
   id,
@@ -21,42 +24,48 @@ export const ListItem = ({
   const secondary = formatArrayAsCommaDelineatedString(
     description,
   );
+  const cls = useStyles({ listItemTextColor: color });
 
   return (
     <ListItemMui
       id={id}
       disableGutters
-      onClick={() => {
-        if (href) navigate(href);
-      }}
-      button={Boolean(href)}
       component="li"
       dense
     >
-      {Icon && (
-        <ListItemAvatarMui>
-          <Avatar style={{ backgroundColor: color }}>
-            <Icon />
-          </Avatar>
-        </ListItemAvatarMui>
-      )}
-      <ListItemTextMui
-        style={{
-          color,
-          textDecoration: href ? 'underline' : 'none',
-        }}
-        primary={primary}
-        secondary={secondary}
-        {...(renderListItemProps
-          ? renderListItemProps()
-          : {})}
-      />
+      <Box
+        component={href ? Link : 'span'}
+        to={href}
+        display="flex"
+        alignItems="center"
+      >
+        {Icon && (
+          <ListItemAvatarMui>
+            <Avatar style={{ backgroundColor: color }}>
+              <Icon />
+            </Avatar>
+          </ListItemAvatarMui>
+        )}
+        <ListItemTextMui
+          style={{
+            color,
+            textDecoration: href ? 'underline' : 'none',
+          }}
+          className={cls.listItemText}
+          primaryTypograph
+          primary={primary}
+          secondary={secondary}
+          {...(renderListItemProps
+            ? renderListItemProps()
+            : {})}
+        />
 
-      {children && !Array.isArray(children)
-        ? React.cloneElement(children, {
-            id,
-          })
-        : null}
+        {children && !Array.isArray(children)
+          ? React.cloneElement(children, {
+              id,
+            })
+          : null}
+      </Box>
     </ListItemMui>
   );
 };
@@ -98,12 +107,16 @@ ListItem.propTypes = {
     PropTypes.string,
     PropTypes.object,
   ]),
+  color: PropTypes.string,
+  href: PropTypes.string,
 };
 
 ListItem.defaultProps = {
   children: null,
   icon: null,
   renderListItemProps: null,
+  color: '',
+  href: '',
 };
 
 export default ListItem;