+++ packages/q3-ui-filemanager/CHANGELOG.md
@@ -3,6 +3,75 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.27.4](https://github.com/3merge/q/compare/v1.27.3...v1.27.4) (2021-02-05)
+
+**Note:** Version bump only for package q3-ui-filemanager
+
+
+
+
+
+## [1.27.3](https://github.com/3merge/q/compare/v1.27.2...v1.27.3) (2021-02-05)
+
+**Note:** Version bump only for package q3-ui-filemanager
+
+
+
+
+
+## [1.27.2](https://github.com/3merge/q/compare/v1.27.1...v1.27.2) (2021-02-04)
+
+**Note:** Version bump only for package q3-ui-filemanager
+
+
+
+
+
+## [1.27.1](https://github.com/3merge/q/compare/v1.27.0...v1.27.1) (2021-02-01)
+
+**Note:** Version bump only for package q3-ui-filemanager
+
+
+
+
+
+# [1.27.0](https://github.com/3merge/q/compare/v1.26.1...v1.27.0) (2021-02-01)
+
+
+### Bug Fixes
+
+* conflicts ([de30b6f](https://github.com/3merge/q/commit/de30b6f63619a56036a96ca27d3d4b2db77dbcb9))
+
+
+
+
+
+
+# [1.26.0](https://github.com/3merge/q/compare/v1.25.6...v1.26.0) (2021-01-27)
+
+**Note:** Version bump only for package q3-ui-filemanager
+
+
+
+
+
+## [1.25.4](https://github.com/3merge/q/compare/v1.25.3...v1.25.4) (2021-01-21)
+
+**Note:** Version bump only for package q3-ui-filemanager
+
+
+
+
+
+## [1.25.3](https://github.com/3merge/q/compare/v1.25.2...v1.25.3) (2021-01-21)
+
+**Note:** Version bump only for package q3-ui-filemanager
+
+
+
+
+
+
 ## [1.24.38](https://github.com/3merge/q/compare/v1.24.37...v1.24.38) (2020-12-19)
 
 **Note:** Version bump only for package q3-ui-filemanager