+++ packages/q3-ui-notifications/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.27.3](https://github.com/3merge/q/compare/v1.27.2...v1.27.3) (2021-02-05)
+
+**Note:** Version bump only for package q3-ui-notifications
+
+
+
+
+
+## [1.27.1](https://github.com/3merge/q/compare/v1.27.0...v1.27.1) (2021-02-01)
+
+**Note:** Version bump only for package q3-ui-notifications
+
+
+
+
+
+
+# [1.26.0](https://github.com/3merge/q/compare/v1.25.6...v1.26.0) (2021-01-27)
+
+
+### Features
+
+* server side events and react charting library ([#285](https://github.com/3merge/q/issues/285)) ([f48225a](https://github.com/3merge/q/commit/f48225adec22bf671cfa2f31f3f14fe99af2d450))
+
+
+
+
+
+
 ## [1.24.33](https://github.com/3merge/q/compare/v1.24.32...v1.24.33) (2020-12-09)